Update SECURITY.md for better readability
It wasn't obvious to me if including the url was on purpose or poor formatting, so I tried my hand at an update that makes it clear one way or another. I personally feel like it is an improvement but also feel like it could be further improved on as well since it still seems kind of off.